requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary
requests==2.27.1
beautifulsoup4==4.10.0
python-dotenv==0.19.2
selenium==4.1.0
webdriver_manager==3.5.2